Adds the holoparasites to the legion loot
In order to add them to the loot pool of the Legion but keeping them off the tendril chests, I've created a subtype of tendril chest, Legion chest. It inherits all from the necropolis chest but name and description are slightly different.

I know this might be a bit overengineered but I believe it will make it easier to add or remove loot in the future.